+++ processor/src/utils/config.utils.ts
@@ -16,6 +16,7 @@ export const readConfiguration = () => {
       projectKey: process.env.CTP_PROJECT_KEY as string,
       scope: process.env.CTP_SCOPE as string,
       region: process.env.CTP_REGION as string,
+      authMode: process.env.AUTHENTICATION_MODE as string,
       sessionAudience: (process.env.CTP_SESSION_AUDIENCE as string) || 'https://mc.europe-west1.gcp.commercetools.com',
       sessionIssuer: (process.env.CTP_SESSION_ISSUER as string) || 'gcp-eu',
     },